requirements.txt
mysql-connector-python==8.0.16
pandas==2.1.3
xarray==2023.11.0
python==3.11.4
